Enable security context by default
The Docker image is already sufficiently advanced in that it creates a
non-root user and its own group. Alpine by default uses the group ID
1000 when adding the first group and the user ID 100.

I needed this because I want to use non-root containers but I also
need to mount the CockroachDB user certificates and those are mounted
using the `podSecurityContext.fsGroup` property value.