Expose liveness and readiness probe configuration in values settings
Different environments and different configuration may have different
expectations for liveness and readiness.

We have experienced that the initial startup using declarative 'git'
backend can take 30s or more on some occasions, which means that the
default liveness probe would kill the container before it could
complete.

By exposing these for users to override and configure, alternative
initialDelay, periodSeconds, and thresholds are able to be tweaked and
refined to support the user's specific expectations.

By default, the same probes that were previously hardcoded are
configured in the defaulti values.yaml.